Add files via upload
An E-mail has taken a prominent place in the lives of information users. Because emails are a faster and more effective manner of communicating, they are employed as a primary and vital mode of information sharing. Email plays a crucial part in communication in both personal and professional lives. Since the previous two decades, the rapid expansion in the number of account users and the volume of mail has resulted in a slew of major difficulties. Users have a tough time recognizing spam emails, which is a huge concern. Spam accounts for about 98 percent of the daily billions of emails exchanged. Spammers have become aware of various email filtering solutions already on the market. As a result, spammers are experimenting with various methods for sending spam or junk mail to a large number of people. Spam pictures and pdfs are one of them. There are now no effective tools available on the market to combat this type of spam email. This study reviews existing email classification systems that may identify emails as spam or ham. 
We require the Data Set for the detection of fraud Emails By using Classification Algorithms and Machine Learning Techniques. This experimentation considers the detection of a phished email as a classification problem and this paper describes the use of machine learning algorithms to classify emails as phished or ham.
So this project is a need for those for everyone who uses email, It will Keep the scammers away and protect valuable money and people from getting scammed